+++ CHANGELOG.rst
@@ -10,10 +10,15 @@ Changed
 =======
 - ``controller`` now holds the dictionary of ``links`` and it can be accessed by other NApps by calling ``self.controller.links``.
 - Each ``Link`` now has a ``threading.Lock`` to perform any change or check on its attributes.
+- Kytos shutdown will now wait for every NApp to shutdown completely.
 
 Fixed
 =====
 - Enhanced docker mongo setup script (rs-init.sh) to properly run on PRIMARY node
+
+Added
+=====
+- Added headers to NApp file response which tells browsers not to cache them.
 
 [2025.1.0] - 2025-04-15
 ***********************